requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_migrate
paho-mqtt
pytz